[MIG] website_snippet_product_category: Migration to version 17.0
TT52399
