+++ website_snippet_product_category/README.rst
@@ -41,10 +41,9 @@ Configuration
 You can select what categories can be shown by the snippet and an image
 for the category:
 
-1. Go to website (backend) > eCommerce > eCommerce Categories
+1. Go to website > eCommerce > eCommerce Categories
 2. Create or Edit one
-3. You can see two new options "Published in product category snippet"
-   and "Category Image"
+3. You can see a new options "Published in product category snippet"
 
 Usage
 =====
