+++ website_snippet_product_category/static/description/index.html
@@ -395,10 +395,9 @@ ul.auto-toc {
 <p>You can select what categories can be shown by the snippet and an image
 for the category:</p>
 <ol class="arabic simple">
-<li>Go to website (backend) &gt; eCommerce &gt; eCommerce Categories</li>
+<li>Go to website &gt; eCommerce &gt; eCommerce Categories</li>
 <li>Create or Edit one</li>
-<li>You can see two new options “Published in product category snippet”
-and “Category Image”</li>
+<li>You can see a new options “Published in product category snippet”</li>
 </ol>
 </div>
 <div class="section" id="usage">
