+++ website_snippet_product_category/templates/snippets.xml
@@ -24,7 +24,7 @@
             t-if="categories and cur_level &lt;= 4"
         >
             <div
-                t-attf-class="categ_tree_level {{'pb-2 main_tree_level text-primary text-uppercase' if cur_level == 1 else ''}} {{cur_level &gt; 2 and 'pl-'+str(min((cur_level-2)*2, 5))}}"
+                t-attf-class="categ_tree_level {{'pb-2 main_tree_level text-primary text-uppercase' if cur_level == 1 else ''}} {{cur_level &gt; 2 and 'ps-'+str(min((cur_level-2)*2, 5))}}"
                 t-att-data-tree-level="cur_level"
                 t-if="category.published_in_product_category_snippet"
             >
@@ -32,7 +32,7 @@
                     <div
                         t-if="category.image_128"
                         t-field="category.image_128"
-                        class="categ_img pr-1"
+                        class="categ_img pe-1"
                         t-options="{'widget': 'image','resize': '24x24' }"
                     />
                     <a
